module for chart configs
